+++ app/Models/RequestErrorLog.php
@@ -35,7 +35,7 @@ class RequestErrorLog extends Model
                 'request_error_log.expected_status_code AS request_error_log_expected_status_code',
                 'request_error_log.returned_status_code AS request_error_log_returned_status_code',
                 'request_error_log.request_uri AS request_error_log_request_uri',
-                //'request_error_log.source AS request_error_log_source',
+                'request_error_log.source AS request_error_log_source',
                 'request_error_log.created_at AS request_error_log_created_at'
             )->
             orderByDesc('created_at')->